+++ relay-server/src/processing/spans/process.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 use std::time::Duration;
 
-use relay_event_normalization::{GeoIpLookup, RequiredMode, SchemaProcessor, eap};
+use relay_event_normalization::{GeoIpLookup, ModelMetadata, RequiredMode, SchemaProcessor, eap};
 use relay_event_schema::processor::{ProcessingState, ValueType, process_value};
 use relay_event_schema::protocol::{Span, SpanId, SpanV2};
 use relay_protocol::Annotated;
@@ -134,10 +134,18 @@ fn parse_and_validate_span_attachment(item: &Item) -> Result<(Option<SpanId>, Ex
 
 /// Normalizes individual spans.
 pub fn normalize(spans: &mut Managed<ExpandedSpans>, geo_lookup: &GeoIpLookup, ctx: Context<'_>) {
+    let model_metadata = ctx.global_config.model_metadata();
     spans.retain_with_context(
         |spans| (&mut spans.spans, &spans.headers),
         |span, headers, _| {
-            normalize_span(&mut span.span, headers, geo_lookup, ctx).inspect_err(|err| {
+            normalize_span(
+                &mut span.span,
+                headers,
+                geo_lookup,
+                model_metadata.as_ref(),
+                ctx,
+            )
+            .inspect_err(|err| {
                 relay_log::debug!("failed to normalize span: {err}");
             })
         },
@@ -148,6 +156,7 @@ fn normalize_span(
     span: &mut Annotated<SpanV2>,
     headers: &EnvelopeHeaders,
     geo_lookup: &GeoIpLookup,
+    model_metadata: Option<&ModelMetadata>,
     ctx: Context<'_>,
 ) -> Result<()> {
     let meta = headers.meta();
@@ -160,7 +169,6 @@ fn normalize_span(
     if let Some(span) = span.value_mut() {
         let dsc = headers.dsc();
         let duration = span_duration(span);
-        let model_costs = ctx.global_config.ai_model_costs.as_ref().ok();
         let allowed_hosts = ctx.global_config.options.http_span_allowed_hosts.as_slice();
 
         validate_timestamps(span)?;
@@ -182,7 +190,7 @@ fn normalize_span(
             eap::normalize_dsc(&mut span.attributes, dsc);
         }
         if ctx.is_processing() {
-            eap::normalize_ai(&mut span.attributes, duration, model_costs);
+            eap::normalize_ai(&mut span.attributes, duration, model_metadata);
         }
         eap::normalize_attribute_values(&mut span.attributes, allowed_hosts);
         eap::write_legacy_attributes(&mut span.attributes);
@@ -848,7 +856,7 @@ mod tests {
             &[],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, ctx).unwrap();
+        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -877,7 +885,7 @@ mod tests {
             &[],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, ctx).unwrap();
+        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -907,7 +915,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, ctx).unwrap();
+        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -932,7 +940,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, ctx).unwrap();
+        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -958,7 +966,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, ctx).unwrap();
+        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
 
         assert_attributes_contains(
             &span,