+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features**:
 
+- Use `ModelMetadata` config with context size and utilization. ([#5814](https://github.com/getsentry/relay/pull/5814))
 - Handle minidump placeholders. ([#5849](https://github.com/getsentry/relay/pull/5849))
 
 **Internal**: